use markdown DSL in RuleSetPagePrinter
