+++ detekt-generator/documentation/comments.md
@@ -4,19 +4,19 @@ This rule set provides rules that address issues in comments and documentation
 of the code.
 
 ## Rules in the `comments` rule set:
+
 ### CommentOverPrivateFunction
 
-TODO: Specify description 
+TODO: Specify description
 
 ### CommentOverPrivateProperty
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UndocumentedPublicClass
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UndocumentedPublicFunction
 
-TODO: Specify description 
-
+TODO: Specify description
