+++ detekt-generator/documentation/complexity.md
@@ -3,47 +3,47 @@
 This rule set contains rules that report complex code.
 
 ## Rules in the `complexity` rule set:
+
 ### LongParameterList
 
-TODO: Specify description 
+TODO: Specify description
 
 ### LongMethod
 
-TODO: Specify description 
+TODO: Specify description
 
 ### LargeClass
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ComplexInterface
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ComplexMethod
 
-TODO: Specify description 
+TODO: Specify description
 
 ### StringLiteralDuplication
 
-TODO: Specify description 
+TODO: Specify description
 
 ### MethodOverloading
 
-TODO: Specify description 
+TODO: Specify description
 
 ### NestedBlockDepth
 
-TODO: Specify description 
+TODO: Specify description
 
 ### TooManyFunctions
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ComplexCondition
 
-TODO: Specify description 
+TODO: Specify description
 
 ### LabeledExpression
 
-TODO: Specify description 
-
+TODO: Specify description
