+++ detekt-generator/documentation/empty-blocks.md
@@ -4,59 +4,59 @@ The empty-blocks ruleset contains rules that will report empty blocks of code
 which should be avoided.
 
 ## Rules in the `empty-blocks` rule set:
+
 ### EmptyCatchBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyClassBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyDefaultConstructor
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyDoWhileBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyElseBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyFinallyBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyForBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyFunctionBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyIfBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyInitBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyKtFile
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptySecondaryConstructor
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyWhenBlock
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EmptyWhileBlock
 
-TODO: Specify description 
-
+TODO: Specify description
