+++ detekt-generator/documentation/exceptions.md
@@ -3,55 +3,55 @@
 Rules in this rule set report issues related to how code throws and handles Exceptions.
 
 ## Rules in the `exceptions` rule set:
+
 ### TooGenericExceptionCaught
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ExceptionRaisedInUnexpectedLocation
 
-TODO: Specify description 
+TODO: Specify description
 
 ### TooGenericExceptionThrown
 
-TODO: Specify description 
+TODO: Specify description
 
 ### NotImplementedDeclaration
 
-TODO: Specify description 
+TODO: Specify description
 
 ### PrintStackTrace
 
-TODO: Specify description 
+TODO: Specify description
 
 ### InstanceOfCheckForException
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ThrowingExceptionsWithoutMessageOrCause
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ReturnFromFinally
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ThrowingExceptionFromFinally
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ThrowingExceptionInMain
 
-TODO: Specify description 
+TODO: Specify description
 
 ### RethrowCaughtException
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ThrowingNewInstanceOfSameException
 
-TODO: Specify description 
+TODO: Specify description
 
 ### SwallowedException
 
-TODO: Specify description 
-
+TODO: Specify description
