+++ detekt-generator/documentation/potential-bugs.md
@@ -3,41 +3,42 @@
 The potential-bugs rule set provides rules that detect potential bugs.
 
 ## Rules in the `potential-bugs` rule set:
+
 ### DuplicateCaseInWhenExpression
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EqualsAlwaysReturnsTrueOrFalse
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EqualsWithHashCodeExist
 
-TODO: Specify description 
+TODO: Specify description
 
 ### IteratorNotThrowingNoSuchElementException
 
-TODO: Specify description 
+TODO: Specify description
 
 ### IteratorHasNextCallsNextMethod
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UselessPostfixExpression
 
-TODO: Specify description 
+TODO: Specify description
 
 ### InvalidLoopCondition
 
-TODO: Specify description 
+TODO: Specify description
 
 ### WrongEqualsTypeParameter
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ExplicitGarbageCollectionCall
 
-TODO: Specify description 
+TODO: Specify description
 
 ### LateinitUsage
 
@@ -46,29 +47,20 @@ Turn on this rule to flag usages of the lateinit modifier.
 Using lateinit for property initialization can be error prone and the actual initialization is not
 guaranteed. Try using constructor injection or delegation to initialize properties.
 
-#### Configuration options: 
-* excludeAnnotatedProperties (default: "")
-
-   Allows you to provide a list of annotations that disable
-this check.
-
-* ignoreOnClassesPattern (default: "")
-
-   Allows you to disable the rule for a list of classes
+#### Configuration options:
 
 ### UnconditionalJumpStatementInLoop
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UnreachableCode
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UnsafeCallOnNullableType
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UnsafeCast
 
-TODO: Specify description 
-
+TODO: Specify description
