+++ detekt-generator/documentation/style.md
@@ -5,9 +5,10 @@ This will help keep code in line with the given
 code style guidelines.
 
 ## Rules in the `style` rule set:
+
 ### CollapsibleIfStatements
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ReturnCount
 
@@ -16,18 +17,15 @@ Restrict the number of return methods allowed in methods.
 Having many exit points in a function can be confusing and impacts readability of the
 code.
 
-#### Configuration options: 
-* max (default: 2)
-
-   define the maximum number of return statements allowed per function
+#### Configuration options:
 
 ### ThrowsCount
 
-TODO: Specify description 
+TODO: Specify description
 
 ### NewLineAtEndOfFile
 
-TODO: Specify description 
+TODO: Specify description
 
 ### WildcardImport
 
@@ -36,95 +34,91 @@ which classes are imported and helps prevent naming conflicts.
 
 Library updates can introduce naming clashes with your own classes which might result in compilation errors.
 
-#### Configuration options: 
-* excludeImports (default: "")
-
-   Define a whitelist of package names that should be allowed to be imported
-with wildcard imports.
+#### Configuration options:
 
 ### MaxLineLength
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EqualsNullCall
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ForbiddenComment
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ForbiddenImport
 
-TODO: Specify description 
+TODO: Specify description
 
 ### FunctionOnlyReturningConstant
 
-TODO: Specify description 
+TODO: Specify description
 
 ### SpacingBetweenPackageAndImports
 
-TODO: Specify description 
+TODO: Specify description
 
 ### LoopWithTooManyJumpStatements
 
-TODO: Specify description 
+TODO: Specify description
 
 ### MethodNameEqualsClassName
 
-TODO: Specify description 
+TODO: Specify description
 
 ### VariableNaming
 
-TODO: Specify description 
+TODO: Specify description
 
 ### VariableMinLength
 
-TODO: Specify description 
+TODO: Specify description
 
 ### VariableMaxLength
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ConstantNaming
 
-TODO: Specify description 
+TODO: Specify description
 
 ### PackageNaming
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ClassNaming
 
-TODO: Specify description 
+TODO: Specify description
 
 ### EnumNaming
 
-TODO: Specify description 
+TODO: Specify description
 
 ### FunctionNaming
 
-TODO: Specify description 
+TODO: Specify description
 
 ### FunctionMaxLength
 
-TODO: Specify description 
+TODO: Specify description
 
 ### FunctionMinLength
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ForbiddenClassName
 
-TODO: Specify description 
+TODO: Specify description
 
 ### SafeCast
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UnnecessaryAbstractClass
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UnnecessaryParentheses
 
@@ -134,39 +128,39 @@ Added in v1.0.0.RC4
 
 ### UnnecessaryInheritance
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UtilityClassWithPublicConstructor
 
-TODO: Specify description 
+TODO: Specify description
 
 ### OptionalAbstractKeyword
 
-TODO: Specify description 
+TODO: Specify description
 
 ### OptionalWhenBraces
 
-TODO: Specify description 
+TODO: Specify description
 
 ### OptionalReturnKeyword
 
-TODO: Specify description 
+TODO: Specify description
 
 ### OptionalUnit
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ProtectedMemberInFinalClass
 
-TODO: Specify description 
+TODO: Specify description
 
 ### SerialVersionUIDInSerializableClass
 
-TODO: Specify description 
+TODO: Specify description
 
 ### MagicNumber
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ModifierOrder
 
@@ -174,25 +168,24 @@ Modifier order array taken from ktlint: https://github.com/shyiko/ktlint
 
 ### DataClassContainsFunctions
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UseDataClass
 
-TODO: Specify description 
+TODO: Specify description
 
 ### UnusedImports
 
-TODO: Specify description 
+TODO: Specify description
 
 ### ExpressionBodySyntax
 
-TODO: Specify description 
+TODO: Specify description
 
 ### NestedClassesVisibility
 
-TODO: Specify description 
+TODO: Specify description
 
 ### RedundantVisibilityModifierRule
 
-TODO: Specify description 
-
+TODO: Specify description
